refactor(mobile): split account form flows
- extract financial account form module\n- split management service planning helpers\n- reduce file size and strict complexity
