+++ apps/mobile/features/financial-accounts/lib/management-service/creation-plan.ts
@@ -1,0 +1,36 @@
+import { buildNewAccountRow } from "./account-rows";
+import { buildManualIdentifierRow } from "./manual-identifier-row";
+import { buildCreateOpeningBalanceRow } from "./opening-balance-rows";
+import { normalizeAccountShape } from "./shape";
+import type { CreateAccountPlan, PlanAccountCreationInput } from "./types";
+
+export function planAccountCreation(input: PlanAccountCreationInput): CreateAccountPlan {
+  const shape = normalizeAccountShape(input.input);
+  const accountId = input.deps.createAccountId();
+  const account = buildNewAccountRow({
+    accountId,
+    userId: input.input.userId,
+    shape,
+    createdAt: input.createdAt,
+  });
+  const openingBalance = buildCreateOpeningBalanceRow({
+    createOpeningBalanceId: input.deps.createOpeningBalanceId,
+    userId: input.input.userId,
+    accountId,
+    openingBalance: shape.openingBalance,
+    createdAt: input.createdAt,
+  });
+  const manualIdentifier = buildManualIdentifierRow({
+    createIdentifierId: input.deps.createIdentifierId,
+    userId: input.input.userId,
+    accountId,
+    value: input.input.manualIdentifierValue,
+    updatedAt: input.createdAt,
+  });
+
+  return {
+    account,
+    openingBalance,
+    manualIdentifier,
+  };
+}
